fix for compartment creation without membranes
Cell and nucleus areas are now compared before compartment creation: if both regions are identical, the compartment creation skips the current pair and the nuclei color is used for selection filling.
Furthermore, "newArray(toString(n))" prevents the creation of an array with n elements, but works with region selection in the ROI Manager.