add feature to select slices without metadata for labeling
Channel selection can now be done with numeral string, i.e. "1", "2", etc. to indicate the slice number, when the metadata for labeling is missing.

Worked around and issue with the "Make Substack..." function that failed to initiate the function argument in the macro, if it's a single numeral string (initiates to an empty string instead).